refactor: beir benchmarking script
