+++ benchmarks/beir-benchmarks/requirements.txt
@@ -1,0 +1,4 @@
+beir>=2.2.0
+llama-stack>=0.2.15
+llama-stack-client>=0.2.15
+pymilvus>=2.5.12
