--- benchmarks/llama-stack-rag-with-beir/requirements.txt
@@ -1,5 +1,0 @@
-beir>=2.2.0
-llama-stack>=0.2.13
-pymilvus>=2.5.12
-pymilvus-model>=0.3.2
-pytrec-eval>=0.5
